+++ src/main/java/com/fatico/winthing/systems/desktop/Module.java
@@ -1,15 +1,12 @@
 package com.fatico.winthing.systems.desktop;
 
-import com.google.inject.PrivateModule;
-import com.google.inject.Singleton;
+import com.google.inject.AbstractModule;
 
-public class Module extends PrivateModule {
+public class Module extends AbstractModule {
 
     @Override
     protected void configure() {
-        bind(DesktopService.class).in(Singleton.class);
         bind(DesktopController.class).asEagerSingleton();
-        expose(DesktopService.class);
     }
 
 }