+++ src/main/java/com/fatico/winthing/systems/monitoring/Module.java
@@ -1,16 +1,12 @@
 package com.fatico.winthing.systems.monitoring;
 
-import com.google.inject.PrivateModule;
-import com.google.inject.Singleton;
+import com.google.inject.AbstractModule;
 
-public class Module extends PrivateModule {
+public class Module extends AbstractModule {
 
     @Override
     protected void configure() {
-        bind(MonitoringService.class).in(Singleton.class);
         bind(MonitoringController.class).asEagerSingleton();
-        expose(MonitoringService.class);
-        expose(MonitoringController.class);
     }
 
 }